requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
aiofiles>=23.2.0
qrcode>=7.4.0
Pillow>=10.0.0